Added a command-line option to store all downloaded subs in a separate subtitles folder. ( --subtitles-folder)
This will still first download the subtitles file in the directory where the media file is located. So it updates the time on that file/directory which is
something we want to avoid if we have scripts processing our media files based on the age of the files.

We have to assume the underlying filesystem doesn't support "time of birth" after stat() call.

See http://unix.stackexchange.com/questions/24441/get-file-created-creation-time